requirements.txt
fastapi==0.95.0
python-dotenv==1.0.0
pymongo==4.3.3
neo4j==5.4.0
pydantic==1.10.5
bson==0.5.10
uvicorn==0.22.0
